Add filtering of non-english streams
The word "Alive" in the top right corner can be localized to a specific
country's primary language if a streamer has their PUBG language set to
something besides english. Since it is right-justified, the position of
the number can change and this will throw off our image cropping and
OCR. For this reason, we will only stick to English streams for the
time being.

Additional notes:

* grab_list_streams has been renamed list_streams
* gtimeout invocations are mac-specific and have been changed to just
  timeout.
* timeout commands have had their multi-line spacing fixed.

Signed-off-by: Cullen Taylor <cullentaylor@outlook.com>