Merge crop_and_ocr.sh into grab_streams.sh
This will take the image cropping/OCR functionality and port it into
grab_streams.sh. Now grab_streams.sh will get a list of english lang
streams from twitch, get a few seconds of video from the stream, make
a single-frame thumbnail for each of those clips, then crop those
thumbnails to just be the "alive" number and interpret its value.
Non-integer values, zero values, and "1" values are ignored as that
means the stream is either not in-game or is at the end of the game
where nothing is happening.

Full notes:
  * add dir_check() to generalize checking existence of dirs.
  * break streams var definition in list_streams() into multiple lines.
  * add some comments about returning non-integer values in functions
    that do so.
  * add get_lowest_stream() to determine the stream closest to the end
    of a game through cropping and tesseract'ing the thumbnails gener-
    ated earlier in the script.
  * add write_index() to write out a very simple index.html with an
    embedded twitch stream. The twitch stream to display is determined
    by get_lowest_stream().
  * add cleanup() function to remove unneeded directories after the
    best stream has been decided.

Signed-off-by: Cullen Taylor <cullentaylor@outlook.com>